backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
chromadb>=0.5
sentence-transformers>=2.7
google-genai>=1.0
pydantic-settings>=2.0
